src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.8.1
    # via slackclient
aiosignal==1.2.0
    # via aiohttp
altgraph==0.17.2
    # via
    #   macholib
    #   pyinstaller
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
charset-normalizer==2.0.12
    # via aiohttp
click==8.1.3
    # via pip-tools
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.3
    # via yarl
macholib==1.16
    # via pyinstaller
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.1
    # via -r requirements.in
pyinstaller==5.1
    # via -r requirements.in
pyinstaller-hooks-contrib==2022.5
    # via pyinstaller
pyyaml==6.0
    # via -r requirements.in
rel==0.4.7
    # via -r requirements.in
slackclient==2.9.4
    # via -r requirements.in
tomli==2.0.1
    # via pep517
websocket-client==1.3.2
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
